rename AppLayout.tsx to AuthenticatedLayout.tsx
